LPD-27493 portal-search-web: Rename date facet configuration js for clarity
