+++ modules/apps/portal-search/portal-search-web/src/main/resources/META-INF/resources/js/shared/RangeInputs.js
