+++ modules/apps/portal-search/portal-search-web/src/main/resources/META-INF/resources/modified/facet/configuration.jsp
@@ -136,5 +136,5 @@ JSONArray rangesJSONArray = modifiedFacetPortletPreferences.getRangesJSONArray()
 			"namespace", liferayPortletResponse.getNamespace()
 		).build()
 	%>'
-	module="{DateFacetConfiguration} from portal-search-web"
+	module="{RangeInputs} from portal-search-web"
 />
